+++ src/main/groovy/ru/vyarus/gradle/plugin/lib/JavaLibPlugin.groovy
@@ -21,6 +21,8 @@ import org.gradle.api.tasks.compile.GroovyCompile
 import org.gradle.api.tasks.compile.JavaCompile
 import org.gradle.api.tasks.javadoc.Javadoc
 import org.gradle.api.tasks.testing.Test
+import org.gradle.plugins.signing.Sign
+import org.gradle.plugins.signing.SigningPlugin
 import org.gradle.process.internal.JvmOptions
 import ru.vyarus.gradle.plugin.pom.PomPlugin
 
@@ -79,6 +81,7 @@ class JavaLibPlugin implements Plugin<Project> {
             MavenPublication bom = configureBomPublication(project)
             configurePlatform(project, extension, bom)
             configureGradleMetadata(project, extension)
+            configureSigning(project, bom)
             addInstallTask(project) {
                 project.logger.warn "INSTALLED $project.group:$bom.artifactId:$project.version"
             }
@@ -96,6 +99,7 @@ class JavaLibPlugin implements Plugin<Project> {
             addJavadocJarTask(project, publication, extension)
             addGroovydocJarTask(project, publication, extension)
             configureGradleMetadata(project, extension)
+            configureSigning(project, publication)
             addInstallTask(project) {
                 project.logger.warn "INSTALLED $project.group:$project.name:$project.version"
             }
@@ -297,6 +301,18 @@ class JavaLibPlugin implements Plugin<Project> {
                 project.tasks.withType(GenerateModuleMetadata).configureEach {
                     enabled = false
                 }
+            }
+        }
+    }
+
+    private void configureSigning(Project project, MavenPublication publication) {
+        project.plugins.withType(SigningPlugin) {
+            // https://docs.gradle.org/current/userguide/signing_plugin.html#sec:signatory_credentials
+            project.signing.sign publication
+
+            // snapshots not signed to simplify project usage (no need for cert during local dev)
+            project.tasks.withType(Sign).configureEach {
+                it.onlyIf { !project.version.toString().endsWith('SNAPSHOT') }
             }
         }
     }