+++ src/test/groovy/ru/vyarus/gradle/plugin/lib/AbstractKitTest.groovy
@@ -35,7 +35,10 @@ abstract class AbstractKitTest extends Specification {
     File fileFromClasspath(String toFile, String source) {
         File target = file(toFile)
         target.parentFile.mkdirs()
-        target << (getClass().getResourceAsStream(source) ?: getClass().classLoader.getResourceAsStream(source)).text
+        target.withOutputStream {
+            it.write((getClass().getResourceAsStream(source) ?: getClass().classLoader.getResourceAsStream(source)).bytes)
+        }
+        target
     }
 
     /**